Small fix in nick readings, saving all to one directory

+++ modules/matching.py
@@ -18,8 +18,8 @@ for pic_path in all_pic_paths:
     "Iterate over pictures to find names in it"
 
     "Print picture and directory"
-    cur_dir = os.path.join(f"exports", os.path.basename(pic_path)[:-4])
-    print("\n", pic_path.ljust(80), cur_dir)
+    cur_dir = os.path.join(f"exports")
+    print(pic_path.ljust(80))
 
     "Load picture"
     img_bgr = cv2.imread(pic_path)
@@ -35,15 +35,13 @@ for pic_path in all_pic_paths:
     os.makedirs(cur_dir, exist_ok=True)
     for num, pt in enumerate(zip(*loc)):
         "Save matched positions"
-        image_path = os.path.join(cur_dir, f"pic_{num:>02}_{int(time.time() * 10000) % 1000}.png")
+        image_path = os.path.join(cur_dir, f"pic_{num:>02}_{int(time.time() * 10000)}.png")
         h, w = pt
-
+        if os.path.isfile(image_path):
+            print(f"This file exists! {image_path}")
+            continue
         pic_of_name = img_bgr[h:h + height, w - extend:w + width]
-        # text = pytesseract.image_to_string(pic_of_name)
         cv2.imwrite(image_path, pic_of_name)
-        # print(f"'{text}'")
-        # cv2.imwrite(os.path.join(cur_dir, text), pic_of_name)
 
 print("Iteration over")
-# cv2.waitKey(0)
 cv2.destroyAllWindows()
